+++ MyVulkan/MyVulkan/FenceGenerator.h
@@ -10,7 +10,7 @@ public:
     FenceGenerator();
     ~FenceGenerator();
 
-    void Create(vk::Device logicalDevice, uint32_t fenceCount);
+    void Create(vk::Device logicalDevice, size_t fenceCount);
     void Destroy();
 
     std::vector<vk::Fence> GetFence();
@@ -19,7 +19,7 @@ private:
     vk::Device m_LogicalDevice;
     std::vector<vk::Fence> m_Fences;
 
-    std::vector<vk::Fence> CreateFences(vk::Device logicalDevice, uint32_t semaphoreCount);
+    std::vector<vk::Fence> CreateFences(vk::Device logicalDevice, size_t semaphoreCount);
 
 };
